Move `SeedSettingsControl` to own file
